[STUDENT MODULES] Corrected component to work on students
