+++ src/components/ModuleCard/ModuleCard.module.css
@@ -1,5 +1,5 @@
 .container {
-    margin: 3rem;
+    margin: 3rem 0;
     transition: all 2s;
 }
 
@@ -7,13 +7,25 @@
     display: grid;
     grid-template-columns: 3fr 3fr 1fr;
     width: 100%;
+    background-color: white;
+    position: relative;
+    z-index: 10;
+    box-shadow: inset;
+}
+
+.topic-card:hover, .challenge-card:hover {
+    transform: scale(1);
 }
 
 .icon {
-    width: 3rem;
-    height: 3rem;
+    width: 4rem;
+    height: 4rem;
     align-self: center;
     justify-self: center;
+}
+
+.icon:hover {
+    cursor: pointer;
 }
 
 .title {
@@ -30,10 +42,29 @@
     justify-items: center;
     align-items: center;
     width: 100%;
+    animation: challenge 1s ease-in-out;
+    box-shadow: none;
+    border-bottom: 0.05rem solid #a1a0a0;
+    border-left: 0.05rem solid rgb(161, 160, 160);
+    border-right: 0.05rem solid rgb(161, 160, 160);
+    position: relative;
+    z-index: 0;
 }
 
 .hidden-card {
     display: none;
+    animation: challenge 1s ease-in-out reverse;
+}
+
+@keyframes challenge {
+    from {
+        transform: translateY(-10rem);
+        opacity: 0;
+    }
+    to {
+        transform: translateY(0);
+        opacity: 1;
+    }
 }
 
 /* SPANS */
@@ -72,10 +103,23 @@
 
 @media (max-width: 30em) {
     .topic-card {
-        grid-template-columns: 1fr;
+        grid-template-columns: 4fr 1fr;
         grid-row-gap: 1rem;
         justify-items: center;
         padding: 2rem 4rem;
+    }
+    .topic-card :first-child {
+        grid-row: 1;
+        grid-column: 1 /3;
+    }
+    .topic-card :nth-child(2) {
+        grid-row: 2;
+        grid-column: 1 ;
+    }
+    .topic-card :last-child {
+        grid-row: 2;
+        grid-column: 2;
+        justify-self: right;
     }
     .challenge-card {
         grid-template-columns: repeat(2, 1fr);
