--- src/components/ModuleStudents/ModuleStudents.module.css
@@ -1,38 +1,0 @@
-.modules {
-    margin: 2rem 0 0 0;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 2rem 4rem;
-}
-
-.title {
-    font-weight: 300;
-    font-size: var(--fs-mid-small-text)
-}
-
-@media (max-width : 60em) {
-    .modules {
-        grid-template-columns: repeat(2, 1fr);
-    }
-}
-
-@media (max-width : 30em) {
-    .modules {
-        grid-template-columns: 1fr;
-    }
-}
-
-.icon-container {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.progress-bar {
-    width: 7rem;
-    height: 7rem;
-}
-
-.icon {
-    width: 6rem;
-    height: 6rem;
-}
