+++ src/components/ModuleTeacherCards/ModuleTeacherCards.module.css
@@ -1,3 +1,18 @@
+.card {
+    min-height: 18rem;
+    padding: 3rem 4rem;
+    border-radius: 1rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    transition: all 500ms;
+}
+
+.locked {
+    background-color: #C6C3C3;
+}
+
+
 .cancel {
     background-color: var(--middle-red);
 }
