--- src/components/ModuleTopics/ModuleTopics.module.css
@@ -1,5 +1,0 @@
-.title {
-    text-align: center;
-    font-size: var(--fs-mid-text);
-    margin: 2rem;
-}
