+++ src/routes/class/Class.tsx
@@ -26,7 +26,9 @@ function ClassNav({ isStudent }: ClassNavProps) {
           </NavLink>
         </li>
         <li>
-          <NavLink to="modules">Módulos</NavLink>
+          <NavLink to={`modules/${isStudent ? 'student' : 'teacher'}`}>
+            Módulos
+          </NavLink>
         </li>
         <li>
           <NavLink to="activities">Actividades</NavLink>
