--- src/routes/modules/modules.module.css
@@ -1,14 +1,0 @@
-
-.card {
-    min-height: 18rem;
-    padding: 3rem 4rem;
-    border-radius: 1rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    transition: all 500ms;
-}
-
-.locked {
-    background-color: #C6C3C3;
-}
